Fix MSVC C4477 warning
Also take translation in to account when printing XP info

+++ Source/qol/common.h
@@ -17,7 +17,7 @@ struct Surface;
  * @param n Number to print
  * @return Address of first character after printed number
  */
-char *PrintWithSeparator(char *out, int64_t n);
+char *PrintWithSeparator(char *out, int n);
 
 void FreeQol();
 void InitQol();
